do not remove the local waf cache if there are more commands to execute
git-svn-id: http://waf.googlecode.com/svn/branches/waf-1.6@8752 f0382ac9-c320-0410-b3f0-b508d59f5a85

+++ waflib/Scripting.py
@@ -281,7 +281,7 @@ def distclean(ctx):
 					Logs.warn('file %r cannot be removed' % f)
 
 		# remove the local waf cache
-		if f.startswith('.waf-'):
+		if f.startswith('.waf-') and not Options.commands:
 			shutil.rmtree(f, ignore_errors=True)
 
 def dist(ctx):
